backend/requirements.txt
flask
google-genai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
praw
vaderSentiment
networkx
scikit-learn
scipy